Go back to pushing Nightlies to the download server, just under a different directory name (and don't update the latest link, since that's just for 9.3-STABLE train releases).

+++ build/post-to-download.sh
@@ -29,10 +29,13 @@ if [ ! -d ${STAGE}/$VERSION-$DATE ]; then
 	exit 2
 fi
 
-TARGET=/tank/downloads/9.3/STABLE/$DATE
+TDIR=`echo ${TRAIN}|awk -F- '{print $2 "/" $3}'`
+TARGET=/tank/downloads/$TDIR/$DATE
 
 ssh ${ID}@download.freenas.org rm -rf $TARGET
 ssh ${ID}@download.freenas.org mkdir -p $TARGET
 scp -pr $STAGE/$VERSION-$DATE/* ${ID}@download.freenas.org:$TARGET
-ssh ${ID}@download.freenas.org "(cd /tank/downloads/${PUSHIT}; rm -f latest; ln -s STABLE/$DATE latest)"
+if [ `echo ${TRAIN}|awk -F- '{print $3}'` != "Nightlies" ]; then
+	ssh ${ID}@download.freenas.org "(cd /tank/downloads/${PUSHIT}; rm -f latest; ln -s STABLE/$DATE latest)"
+fi
 exit 0
